stopping language runner if not nil. stopping runner only at shutdown request

+++ api/lang/runner.go
@@ -110,5 +110,8 @@ func getAllStepsResponse() (*gm.StepNamesResponse, error) {
 }
 
 func killRunner() {
-	lRunner.runner.Kill()
+	if lRunner.runner != nil {
+		lRunner.runner.Kill()
+	}
+
 }
